requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.1
django-oauth-toolkit==2.4.0
django-oidc-provider==0.8.2
djangorestframework==3.15.2
future==1.0.0
gunicorn==23.0.0
idna==3.7
jwcrypto==1.5.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
pyjwkest==1.4.2
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
